+++ spec/grape/endpoint_spec.rb
@@ -1021,12 +1021,12 @@ describe Grape::Endpoint do
     end
 
     it 'result in a 406 response if they are invalid' do
-      get '/test', {}, Grape::Http::Headers::HTTP_ACCEPT => 'application/vnd.ohanapi.v1+json'
+      get '/test', {}, 'HTTP_ACCEPT' => 'application/vnd.ohanapi.v1+json'
       expect(last_response.status).to eq(406)
     end
 
     it 'result in a 406 response if they cannot be parsed' do
-      get '/test', {}, Grape::Http::Headers::HTTP_ACCEPT => 'application/vnd.ohanapi.v1+json; version=1'
+      get '/test', {}, 'HTTP_ACCEPT' => 'application/vnd.ohanapi.v1+json; version=1'
       expect(last_response.status).to eq(406)
     end
   end